tictaktoe_QA/synth_dataset/requirements.txt
datasets>=2.20.0
numpy>=1.24
pillow>=10.0
